+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2025-10-08
+
 ### Changed
 
 - Use api version `v2` when querying Helm custom resources.
@@ -793,7 +795,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added initial framework layout
 - Added Kubernetes client extended from controller-runtime client
 
-[Unreleased]: https://github.com/giantswarm/clustertest/compare/v1.41.0...HEAD
+[Unreleased]: https://github.com/giantswarm/clustertest/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/giantswarm/clustertest/compare/v1.41.0...v2.0.0
 [1.41.0]: https://github.com/giantswarm/clustertest/compare/v1.40.0...v1.41.0
 [1.40.0]: https://github.com/giantswarm/clustertest/compare/v1.39.3...v1.40.0
 [1.39.3]: https://github.com/giantswarm/clustertest/compare/v1.39.2...v1.39.3